Move MinMaxSource to separate file
